use only phrase search logic

+++ src/main/kotlin/com/noodle/upper/models/Tracked.kt
@@ -1,6 +1,6 @@
 package com.noodle.upper.models
 
-data class Tracked<T>(val fetched: Int, val total: Int, val entity: T?) {
+data class Tracked<T>(val fetched: Int, val total: Long, val entity: T?) {
 }
 
 fun <T> Tracked<T>.completion(): Float = this.fetched.toFloat()/this.total
